[resumes][refactor] combine resume-card and resume-body into resume-list-item
